copy and don't delete node.csv
git-svn-id: file:///srv/local/data/cleanRepo/trunk@22003 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/mrieser/src/main/java/playground/mrieser/svi/controller/DynusTControlerListener.java
@@ -132,7 +132,7 @@ public class DynusTControlerListener implements StartupListener, IterationStarts
 			log.info("removing unnecessary files.");
 			File outDir = new File(this.dc.getOutputDirectory());
 			for (File f : outDir.listFiles()) {
-				if (f.isFile() && !f.getName().equals("VehTrajectory.dat")) {
+				if (f.isFile() && !f.getName().equals("VehTrajectory.dat") && !f.getName().equalsIgnoreCase("node.csv")) {
 					log.info("   deleting file " + f.getAbsolutePath());
 					f.delete();
 				}
